MDL-48771 quiz: Select multiple questions to delete
Upon Moodle 2.8 the option to bulk delete all questions of a quiz
disappeared.
This patch adresses this and reintruduces a bulk action button.
Pushing that button shows a Delete and a Cancel button as well as
two Select all / Deselect all triggers and each and every question
gets a checkbox. The UI is slimmed down and other elements (e.g.
'Repaginate' button and so on) are hidden. Questions can be chosen
and deleted by pushing the Delete button. The bulk action button
is disabled when quiz attempts are present, just like the
'Repaginate' button.

TODO: Behat tests are lacking.