MDL-61917 question regrade: safer handling of manual grading
Questions can only be manually graded after they are finished. With some
behaviours, if the question is edited in certain ways, when the question
is regraded there may be a point where it is not finished in the new
attempt, when it was in the old attempt.

This patch ensure that the question is finished before a manual grading
action is replayed.