MDL-69704 component_callback_exists: uninstalled plugin not an error
component_callback_exists was considering it a fatal error to call the
function with the name of a component that had been uninstalled. This
seems unduly harsh, and was breaking things like the activity chooser.
The activity chooser was getting a list of all subplugins (using
$subplugins = $pluginmanager->get_subplugins_of_plugin('mod_' . $plugin->name);
which seems to include 'missing from disc' plugins, and then calling
component_callback_exists which was throwing an exception, and so
causing serious breakage.

Returning false instead seems to be a much safer option, and seems
to be much more in the spirit of what the component_callback_exists
API is trying to do.