MDL-75576 quiz statistics: don't compute when a user views the qbank
Previously, when a users viewed the question bank, if the quiz statistics
had not already been calculated, we would try to compute them there an then,
which could be very, very slow, leading to session lock problems.

Now, we never try to compute the statistics on the fly. Instead, we rely
on the existing \quiz_statistics\task\recalculate scheduled task to do it.