MDL-83862 question: Improve error handling of question bank filters.
This allows filter fields to use client-side form validation attributes to check for valid values, and adds a validation() method for filters to implement custom validation conditions.

In the event of a validation failure, the invalid filters will not be applied, and the URL will not be updated to avoid user sharing a URL that will generate an error. Filter conditions should still implement server-side validation and throw exceptions to avoid erroneous values being manually inserted into the URL.